+++ Runtime/Cryptographers/CryptographerType.cs
@@ -1,8 +1,0 @@
-namespace ActionCode.Persistence
-{
-    public enum CryptographerType
-    {
-        None = -1,
-        AES
-    }
-}